refact: refactored the package into multiple specialized sub-packages
This clarifies the dependencies between the various utilities exposed by
this package.

Ultimately (not in this PR), we may convert some of the sub-packages
into go modules, so as to isolate dependencies and allow users of
some of the features not to be polluted by cross-dependencies.

* refactored mangling to remove package state
