+++ jsonname/name_provider_test.go
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package swag
+package jsonname
 
 import (
 	"reflect"
